minor-backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
motor==3.1.1
pydantic[email]==1.10.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
pymongo==4.6.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
nltk==3.8.1
tensorflow==2.16.1
websockets==10.4
